# TreasureDeckApi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TreasureDeck API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v2/cardsVersion/{id}": {
            "get": {
                "tags": [
                    "cardsVersions"
                ],
                "summary": "Obtener una versión por ID",
                "operationId": "bdffad542b87a1b9bb0e21e0c75856b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la versión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Versión encontrada"
                    },
                    "404": {
                        "description": "Versión no encontrada"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "cardsVersions"
                ],
                "summary": "Actualizar una versión de carta",
                "operationId": "6261599c4dd1cd8c4f249adb699ce41b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la versión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "image_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/card.png"
                                    },
                                    "min_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1.2
                                    },
                                    "avg_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Versión actualizada"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "cardsVersions"
                ],
                "summary": "Eliminar una versión de carta",
                "operationId": "04c6a54fb9e86336ebfddbca652a173f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la versión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Versión eliminada"
                    },
                    "404": {
                        "description": "Versión no encontrada"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/deckCards": {
            "get": {
                "tags": [
                    "deckCards"
                ],
                "summary": "Listar todas las cartas en mazos",
                "operationId": "cefb8f38e2f187a5304df054479dc709",
                "responses": {
                    "200": {
                        "description": "Lista de cartas en mazos"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "deckCards"
                ],
                "summary": "Agregar una carta a un mazo",
                "operationId": "5c345c84d0759bd10e2c9609d8b84ae9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "deck_id",
                                    "card_version_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "deck_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "card_version_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Carta agregada al mazo"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/deckCards/{id}": {
            "get": {
                "tags": [
                    "deckCards"
                ],
                "summary": "Mostrar una carta específica de un mazo",
                "operationId": "fb243983a57840cea54aff4ce42a741d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la carta en el mazo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carta encontrada"
                    },
                    "404": {
                        "description": "Carta no encontrada"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "deckCards"
                ],
                "summary": "Actualizar una carta en el mazo",
                "operationId": "99ed2ad841b1c958449df69ea22b368f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la carta en el mazo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Carta actualizada"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "deckCards"
                ],
                "summary": "Eliminar una carta del mazo",
                "operationId": "1b7721ae4f78243d89b810bd98fda0a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la carta en el mazo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Carta eliminada del mazo"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Autenticación usando Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}